Repository cleanup
* Mention pprint in README.md

* Add cmake generated files to .gitignore

* Move the original Makefile and vgrind.sh off to a directory named
ATTIC.  They are useful enough during development (still!) that I want
to keep them in the HEAD, but also a source of confusion.  A BSD
Makefile in the main project directory is an invitation to mischief.
It can be accidentally run, creating more files in the main source
directory, and might be trusted when it should not be.  Likewise,
vgrind.sh, which no longer works, may be run by accident, creating
mayhem.